move postgres from mango-feeds-lib
